requirements.txt
torch>=2.4,<2.5
numpy>=1.26
pandas>=2.0
scipy>=1.11
matplotlib>=3.8
PyYAML>=6.0
h5py>=3.10
tqdm>=4.66
